Fix user doc creation to respect Firestore rules
